@@ -76,32 +73,31 @@
     },
     "configFile": "system.config.js",
     "dependencies": {
-      "@angular/forms": "npm:@angular/forms@^2.1.0",
-      "@angular/http": "npm:@angular/http@^2.1.0",
-      "@angular/platform-browser-dynamic": "npm:@angular/platform-browser-dynamic@^2.1.0",
-      "@angular/router": "npm:@angular/router@^3.1.0",
-      "angular2-materialize": "npm:angular2-materialize@^6.0.1",
+      "@angular/forms": "npm:@angular/forms@^4.1.1",
+      "@angular/http": "npm:@angular/http@^4.1.1",
+      "@angular/platform-browser-dynamic": "npm:@angular/platform-browser-dynamic@^4.1.1",
+      "@angular/router": "npm:@angular/router@^4.1.1",
+      "angular2-materialize": "npm:angular2-materialize@^15.0.1",
       "animate.css": "npm:animate.css@^3.5.1",
       "css-animator": "npm:css-animator@^2.0.1",
       "flat-color-icons": "npm:flat-color-icons@^1.0.0",
       "iscroll": "npm:iscroll@^5.2.0",
-      "jquery": "npm:jquery@^2.2.4",
-      "materialize": "github:Dogfalo/materialize@^0.97.6",
+      "jquery": "npm:jquery@2",
       "reflect-metadata": "npm:reflect-metadata@^0.1.3",
-      "typescript": "npm:typescript@^1.8.10"
+      "typescript": "npm:typescript@^2.0.0"
     },
     "devDependencies": {
-      "core-js": "npm:core-js@^1.2.0",
-      "ts-runtime": "npm:babel-runtime@^5.8.24",
-      "text": "github:systemjs/plugin-text@^0.0.8",
-      "ts": "github:frankwallis/plugin-typescript@^4.0.16",
-      "plugin-babel": "npm:systemjs-plugin-babel@^0.0.11"
+      "core-js": "npm:core-js@^2.4.1",
+      "ts-runtime": "npm:babel-runtime@^6.23.0",
+      "text": "github:systemjs/plugin-text@^0.0.9",
+      "ts": "github:frankwallis/plugin-typescript@^7.0.6",
+      "plugin-babel": "npm:systemjs-plugin-babel@^0.0.21"
     },
     "peerDependencies": {
-      "@angular/common": "npm:@angular/common@2.4.9",
-      "@angular/compiler": "npm:@angular/compiler@2.4.9",
-      "@angular/core": "npm:@angular/core@2.4.9",
-      "@angular/platform-browser": "npm:@angular/platform-browser@2.4.9",
+      "@angular/common": "npm:@angular/common@4.1.1",
+      "@angular/compiler": "npm:@angular/compiler@4.1.1",
+      "@angular/core": "npm:@angular/core@4.1.1",
+      "@angular/platform-browser": "npm:@angular/platform-browser@4.1.1",
       "assert": "npm:jspm-nodelibs-assert@^0.2.0",
       "buffer": "npm:jspm-nodelibs-buffer@^0.2.0",
       "child_process": "npm:jspm-nodelibs-child_process@^0.2.0",
