@@ -3,7 +3,10 @@ import 'rxjs/add/operator/timeInterval';
 import 'rxjs/add/operator/take';
 
 import { Component, OnInit, Input, Output, EventEmitter, ElementRef } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+
+// import { Observable } from 'rxjs/Observable' => Observable undefined
+import RxObservable from 'rxjs/Observable';
+const Observable = RxObservable.Observable;
 
 import { AnimationService, AnimationBuilder } from 'css-animator';
 import { Unsubscriber } from 'app/components';
