@@ -1,5 +1,8 @@
 import { OnDestroy } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+
+// import { Observable } from 'rxjs/Observable' => Observable undefined
+import RxObservable from 'rxjs/Observable';
+const Observable = RxObservable.Observable;
 
 export class Unsubscriber implements OnDestroy {
 
