@@ -4,7 +4,10 @@ import 'rxjs/add/operator/map';
 
 import { Injectable, EventEmitter, isDevMode } from '@angular/core';
 import { Http, Response } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
+
+// import { Observable } from 'rxjs/Observable' => Observable undefined
+import RxObservable from 'rxjs/Observable';
+const Observable = RxObservable.Observable;
 
 import { shuffle } from 'app/helpers';
 import { TrackTransformer } from 'app/support';
