@@ -9,19 +9,17 @@ SystemJS.config({
   },
   devConfig: {
     "map": {
-      "ts-runtime": "npm:babel-runtime@5.8.38",
-      "core-js": "npm:core-js@1.2.7",
-      "plugin-babel": "npm:systemjs-plugin-babel@0.0.11",
-      "text": "github:systemjs/plugin-text@0.0.8",
-      "ts": "github:frankwallis/plugin-typescript@4.0.16"
+      "ts-runtime": "npm:babel-runtime@6.23.0",
+      "core-js": "npm:core-js@2.4.1",
+      "plugin-babel": "npm:systemjs-plugin-babel@0.0.21",
+      "text": "github:systemjs/plugin-text@0.0.9",
+      "ts": "github:frankwallis/plugin-typescript@7.0.6"
     },
     "packages": {
-      "npm:babel-runtime@5.8.38": {
-        "map": {}
-      },
-      "github:frankwallis/plugin-typescript@4.0.16": {
+      "npm:babel-runtime@6.23.0": {
         "map": {
-          "typescript": "npm:typescript@1.8.10"
+          "core-js": "npm:core-js@2.4.1",
+          "regenerator-runtime": "npm:regenerator-runtime@0.10.5"
         }
       }
     }
@@ -75,46 +73,46 @@ SystemJS.config({
     "github:*/*.json"
   ],
   map: {
-    "@angular/common": "npm:@angular/common@2.4.9",
-    "@angular/compiler": "npm:@angular/compiler@2.4.9",
-    "@angular/core": "npm:@angular/core@2.4.9",
-    "@angular/forms": "npm:@angular/forms@2.4.9",
-    "@angular/http": "npm:@angular/http@2.4.9",
-    "@angular/platform-browser": "npm:@angular/platform-browser@2.4.9",
-    "@angular/platform-browser-dynamic": "npm:@angular/platform-browser-dynamic@2.4.9",
-    "@angular/router": "npm:@angular/router@3.4.9",
-    "angular2-materialize": "npm:angular2-materialize@6.7.1",
+    "@angular/common": "npm:@angular/common@4.1.1",
+    "@angular/compiler": "npm:@angular/compiler@4.1.1",
+    "@angular/core": "npm:@angular/core@4.1.1",
+    "@angular/forms": "npm:@angular/forms@4.1.1",
+    "@angular/http": "npm:@angular/http@4.1.1",
+    "@angular/platform-browser": "npm:@angular/platform-browser@4.1.1",
+    "@angular/platform-browser-dynamic": "npm:@angular/platform-browser-dynamic@4.1.1",
+    "@angular/router": "npm:@angular/router@4.1.1",
+    "angular2-materialize": "npm:angular2-materialize@15.0.1",
     "animate.css": "npm:animate.css@3.5.2",
-    "assert": "npm:jspm-nodelibs-assert@0.2.0",
-    "buffer": "npm:jspm-nodelibs-buffer@0.2.1",
-    "child_process": "npm:jspm-nodelibs-child_process@0.2.0",
-    "constants": "npm:jspm-nodelibs-constants@0.2.0",
-    "crypto": "npm:jspm-nodelibs-crypto@0.2.0",
+    "assert": "npm:jspm-nodelibs-assert@0.2.1",
+    "buffer": "npm:jspm-nodelibs-buffer@0.2.3",
+    "child_process": "npm:jspm-nodelibs-child_process@0.2.1",
+    "constants": "npm:jspm-nodelibs-constants@0.2.1",
+    "crypto": "npm:jspm-nodelibs-crypto@0.2.1",
     "css-animator": "npm:css-animator@2.0.2",
-    "events": "npm:jspm-nodelibs-events@0.2.0",
-    "flat-color-icons": "npm:flat-color-icons@1.0.0",
-    "fs": "npm:jspm-nodelibs-fs@0.2.0",
+    "events": "npm:jspm-nodelibs-events@0.2.2",
+    "flat-color-icons": "npm:flat-color-icons@1.1.0",
+    "fs": "npm:jspm-nodelibs-fs@0.2.1",
     "http": "npm:jspm-nodelibs-http@0.2.0",
     "iscroll": "npm:iscroll@5.2.0",
     "jquery": "npm:jquery@2.2.4",
-    "materialize": "github:Dogfalo/materialize@0.97.8",
-    "materialize-css": "npm:materialize-css@0.97.8",
-    "module": "npm:jspm-nodelibs-module@0.2.0",
-    "net": "npm:jspm-nodelibs-net@0.2.0",
-    "os": "npm:jspm-nodelibs-os@0.2.0",
-    "path": "npm:jspm-nodelibs-path@0.2.1",
-    "process": "npm:jspm-nodelibs-process@0.2.0",
+    "materialize-css": "npm:materialize-css@0.98.2",
+    "module": "npm:jspm-nodelibs-module@0.2.1",
+    "net": "npm:jspm-nodelibs-net@0.2.1",
+    "os": "npm:jspm-nodelibs-os@0.2.1",
+    "path": "npm:jspm-nodelibs-path@0.2.3",
+    "process": "npm:jspm-nodelibs-process@0.2.1",
     "reflect-metadata": "npm:reflect-metadata@0.1.10",
-    "rxjs": "npm:rxjs@5.2.0",
-    "stream": "npm:jspm-nodelibs-stream@0.2.0",
-    "string_decoder": "npm:jspm-nodelibs-string_decoder@0.2.0",
-    "timers": "npm:jspm-nodelibs-timers@0.2.0",
-    "tty": "npm:jspm-nodelibs-tty@0.2.0",
-    "typescript": "npm:typescript@1.8.10",
-    "url": "npm:jspm-nodelibs-url@0.2.0",
-    "util": "npm:jspm-nodelibs-util@0.2.1",
-    "vm": "npm:jspm-nodelibs-vm@0.2.0",
-    "zone.js": "npm:zone.js@0.7.7"
+    "rxjs": "npm:rxjs@5.3.1",
+    "stream": "npm:jspm-nodelibs-stream@0.2.1",
+    "string_decoder": "npm:jspm-nodelibs-string_decoder@0.2.1",
+    "timers": "npm:jspm-nodelibs-timers@0.2.1",
+    "tty": "npm:jspm-nodelibs-tty@0.2.1",
+    "typescript": "npm:typescript@2.3.2",
+    "url": "npm:jspm-nodelibs-url@0.2.1",
+    "util": "npm:jspm-nodelibs-util@0.2.2",
+    "vm": "npm:jspm-nodelibs-vm@0.2.1",
+    "zlib": "npm:jspm-nodelibs-zlib@0.2.3",
+    "zone.js": "npm:zone.js@0.8.10"
   },
   packages: {
     "npm:browserify-aes@1.0.6": {
@@ -144,17 +142,6 @@ SystemJS.config({
       "map": {
         "bn.js": "npm:bn.js@4.11.6",
         "randombytes": "npm:randombytes@2.0.3"
-      }
-    },
-    "npm:browserify-sign@4.0.0": {
-      "map": {
-        "bn.js": "npm:bn.js@4.11.6",
-        "browserify-rsa": "npm:browserify-rsa@4.0.1",
-        "create-hash": "npm:create-hash@1.1.2",
-        "create-hmac": "npm:create-hmac@1.1.4",
-        "elliptic": "npm:elliptic@6.4.0",
-        "inherits": "npm:inherits@2.0.3",
-        "parse-asn1": "npm:parse-asn1@5.0.0"
       }
     },
     "npm:buffer-shims@1.0.0": {
@@ -189,7 +176,7 @@ SystemJS.config({
     "npm:crypto-browserify@3.11.0": {
       "map": {
         "browserify-cipher": "npm:browserify-cipher@1.0.0",
-        "browserify-sign": "npm:browserify-sign@4.0.0",
+        "browserify-sign": "npm:browserify-sign@4.0.4",
         "create-ecdh": "npm:create-ecdh@4.0.0",
         "create-hash": "npm:create-hash@1.1.2",
         "create-hmac": "npm:create-hmac@1.1.4",
@@ -229,21 +216,12 @@ SystemJS.config({
         "brorand": "npm:brorand@1.1.0"
       }
     },
-    "npm:parse-asn1@5.0.0": {
-      "map": {
-        "asn1.js": "npm:asn1.js@4.9.1",
-        "browserify-aes": "npm:browserify-aes@1.0.6",
-        "create-hash": "npm:create-hash@1.1.2",
-        "evp_bytestokey": "npm:evp_bytestokey@1.0.0",
-        "pbkdf2": "npm:pbkdf2@3.0.9"
-      }
-    },
     "npm:public-encrypt@4.0.0": {
       "map": {
         "bn.js": "npm:bn.js@4.11.6",
         "browserify-rsa": "npm:browserify-rsa@4.0.1",
         "create-hash": "npm:create-hash@1.1.2",
-        "parse-asn1": "npm:parse-asn1@5.0.0",
+        "parse-asn1": "npm:parse-asn1@5.1.0",
         "randombytes": "npm:randombytes@2.0.3"
       }
     },
@@ -259,20 +237,13 @@ SystemJS.config({
     "npm:stream-browserify@2.0.1": {
       "map": {
         "inherits": "npm:inherits@2.0.3",
-        "readable-stream": "npm:readable-stream@2.2.3"
+        "readable-stream": "npm:readable-stream@2.2.9"
       }
     },
     "npm:url@0.11.0": {
       "map": {
         "punycode": "npm:punycode@1.3.2",
         "querystring": "npm:querystring@0.2.0"
-      }
-    },
-    "npm:buffer@4.9.1": {
-      "map": {
-        "ieee754": "npm:ieee754@1.1.8",
-        "isarray": "npm:isarray@1.0.0",
-        "base64-js": "npm:base64-js@1.2.0"
       }
     },
     "npm:cipher-base@1.0.3": {
@@ -282,7 +253,7 @@ SystemJS.config({
     },
     "npm:timers-browserify@1.4.2": {
       "map": {
-        "process": "npm:process@0.11.9"
+        "process": "npm:process@0.11.10"
       }
     },
     "npm:pbkdf2@3.0.9": {
@@ -290,84 +261,14 @@ SystemJS.config({
         "create-hmac": "npm:create-hmac@1.1.4"
       }
     },
-    "npm:materialize-css@0.97.8": {
-      "map": {
-        "css": "github:systemjs/plugin-css@0.1.32",
-        "jquery": "github:components/jquery@3.1.1"
-      }
-    },
-    "github:Dogfalo/materialize@0.97.8": {
-      "map": {
-        "jquery": "npm:jquery@2.2.4",
-        "css": "github:systemjs/plugin-css@0.1.32"
-      }
-    },
-    "npm:typescript@1.8.10": {
-      "map": {}
-    },
     "npm:sha.js@2.4.8": {
       "map": {
         "inherits": "npm:inherits@2.0.3"
       }
     },
-    "npm:jspm-nodelibs-os@0.2.0": {
-      "map": {
-        "os-browserify": "npm:os-browserify@0.2.1"
-      }
-    },
-    "npm:jspm-nodelibs-crypto@0.2.0": {
-      "map": {
-        "crypto-browserify": "npm:crypto-browserify@3.11.0"
-      }
-    },
-    "npm:jspm-nodelibs-string_decoder@0.2.0": {
-      "map": {
-        "string_decoder-browserify": "npm:string_decoder@0.10.31"
-      }
-    },
-    "npm:jspm-nodelibs-stream@0.2.0": {
-      "map": {
-        "stream-browserify": "npm:stream-browserify@2.0.1"
-      }
-    },
     "npm:jspm-nodelibs-http@0.2.0": {
       "map": {
-        "http-browserify": "npm:stream-http@2.6.3"
-      }
-    },
-    "npm:jspm-nodelibs-url@0.2.0": {
-      "map": {
-        "url-browserify": "npm:url@0.11.0"
-      }
-    },
-    "npm:jspm-nodelibs-timers@0.2.0": {
-      "map": {
-        "timers-browserify": "npm:timers-browserify@1.4.2"
-      }
-    },
-    "npm:jspm-nodelibs-buffer@0.2.1": {
-      "map": {
-        "buffer": "npm:buffer@4.9.1"
-      }
-    },
-    "npm:stream-http@2.6.3": {
-      "map": {
-        "builtin-status-codes": "npm:builtin-status-codes@3.0.0",
-        "readable-stream": "npm:readable-stream@2.2.3",
-        "xtend": "npm:xtend@4.0.1",
-        "inherits": "npm:inherits@2.0.3",
-        "to-arraybuffer": "npm:to-arraybuffer@1.0.1"
-      }
-    },
-    "npm:readable-stream@2.2.3": {
-      "map": {
-        "inherits": "npm:inherits@2.0.3",
-        "string_decoder": "npm:string_decoder@0.10.31",
-        "isarray": "npm:isarray@1.0.0",
-        "core-util-is": "npm:core-util-is@1.0.2",
-        "buffer-shims": "npm:buffer-shims@1.0.0",
-        "util-deprecate": "npm:util-deprecate@1.0.2",
-        "process-nextick-args": "npm:process-nextick-args@1.0.7"
+        "http-browserify": "npm:stream-http@2.7.0"
       }
     },
     "npm:elliptic@6.4.0": {
@@ -378,14 +279,7 @@ SystemJS.config({
         "minimalistic-assert": "npm:minimalistic-assert@1.0.0",
         "brorand": "npm:brorand@1.1.0",
         "minimalistic-crypto-utils": "npm:minimalistic-crypto-utils@1.0.1",
-        "hmac-drbg": "npm:hmac-drbg@1.0.0"
-      }
-    },
-    "npm:hmac-drbg@1.0.0": {
-      "map": {
-        "hash.js": "npm:hash.js@1.0.3",
-        "minimalistic-assert": "npm:minimalistic-assert@1.0.0",
-        "minimalistic-crypto-utils": "npm:minimalistic-crypto-utils@1.0.1"
+        "hmac-drbg": "npm:hmac-drbg@1.0.1"
       }
     },
     "npm:asn1.js@4.9.1": {
@@ -395,9 +289,198 @@ SystemJS.config({
         "minimalistic-assert": "npm:minimalistic-assert@1.0.0"
       }
     },
-    "npm:rxjs@5.2.0": {
+    "npm:browserify-sign@4.0.4": {
+      "map": {
+        "create-hmac": "npm:create-hmac@1.1.4",
+        "parse-asn1": "npm:parse-asn1@5.1.0",
+        "elliptic": "npm:elliptic@6.4.0",
+        "bn.js": "npm:bn.js@4.11.6",
+        "create-hash": "npm:create-hash@1.1.2",
+        "browserify-rsa": "npm:browserify-rsa@4.0.1",
+        "inherits": "npm:inherits@2.0.3"
+      }
+    },
+    "npm:parse-asn1@5.1.0": {
+      "map": {
+        "browserify-aes": "npm:browserify-aes@1.0.6",
+        "create-hash": "npm:create-hash@1.1.2",
+        "evp_bytestokey": "npm:evp_bytestokey@1.0.0",
+        "pbkdf2": "npm:pbkdf2@3.0.9",
+        "asn1.js": "npm:asn1.js@4.9.1"
+      }
+    },
+    "npm:hmac-drbg@1.0.1": {
+      "map": {
+        "hash.js": "npm:hash.js@1.0.3",
+        "minimalistic-assert": "npm:minimalistic-assert@1.0.0",
+        "minimalistic-crypto-utils": "npm:minimalistic-crypto-utils@1.0.1"
+      }
+    },
+    "npm:readable-stream@2.2.9": {
+      "map": {
+        "isarray": "npm:isarray@1.0.0",
+        "inherits": "npm:inherits@2.0.3",
+        "string_decoder": "npm:string_decoder@1.0.0",
+        "core-util-is": "npm:core-util-is@1.0.2",
+        "buffer-shims": "npm:buffer-shims@1.0.0",
+        "process-nextick-args": "npm:process-nextick-args@1.0.7",
+        "util-deprecate": "npm:util-deprecate@1.0.2"
+      }
+    },
+    "npm:string_decoder@1.0.0": {
+      "map": {
+        "buffer-shims": "npm:buffer-shims@1.0.0"
+      }
+    },
+    "npm:materialize-css@0.98.2": {
+      "map": {
+        "hammerjs": "npm:hammerjs@2.0.8",
+        "node-archiver": "npm:node-archiver@0.3.0",
+        "jquery": "npm:jquery@2.2.4"
+      }
+    },
+    "npm:node-archiver@0.3.0": {
+      "map": {
+        "fstream": "npm:fstream@1.0.11",
+        "tar": "npm:tar@2.2.1"
+      }
+    },
+    "npm:tar@2.2.1": {
+      "map": {
+        "fstream": "npm:fstream@1.0.11",
+        "inherits": "npm:inherits@2.0.3",
+        "block-stream": "npm:block-stream@0.0.9"
+      }
+    },
+    "npm:fstream@1.0.11": {
+      "map": {
+        "mkdirp": "npm:mkdirp@0.5.1",
+        "graceful-fs": "npm:graceful-fs@4.1.11",
+        "rimraf": "npm:rimraf@2.6.1",
+        "inherits": "npm:inherits@2.0.3"
+      }
+    },
+    "npm:jspm-nodelibs-zlib@0.2.3": {
+      "map": {
+        "browserify-zlib": "npm:browserify-zlib@0.1.4"
+      }
+    },
+    "npm:block-stream@0.0.9": {
+      "map": {
+        "inherits": "npm:inherits@2.0.3"
+      }
+    },
+    "npm:mkdirp@0.5.1": {
+      "map": {
+        "minimist": "npm:minimist@0.0.8"
+      }
+    },
+    "npm:browserify-zlib@0.1.4": {
+      "map": {
+        "readable-stream": "npm:readable-stream@2.2.9",
+        "pako": "npm:pako@0.2.9"
+      }
+    },
+    "npm:rimraf@2.6.1": {
+      "map": {
+        "glob": "npm:glob@7.1.1"
+      }
+    },
+    "npm:glob@7.1.1": {
+      "map": {
+        "inherits": "npm:inherits@2.0.3",
+        "fs.realpath": "npm:fs.realpath@1.0.0",
+        "inflight": "npm:inflight@1.0.6",
+        "once": "npm:once@1.4.0",
+        "minimatch": "npm:minimatch@3.0.3",
+        "path-is-absolute": "npm:path-is-absolute@1.0.1"
+      }
+    },
+    "npm:inflight@1.0.6": {
+      "map": {
+        "once": "npm:once@1.4.0",
+        "wrappy": "npm:wrappy@1.0.2"
+      }
+    },
+    "npm:once@1.4.0": {
+      "map": {
+        "wrappy": "npm:wrappy@1.0.2"
+      }
+    },
+    "npm:minimatch@3.0.3": {
+      "map": {
+        "brace-expansion": "npm:brace-expansion@1.1.7"
+      }
+    },
+    "npm:brace-expansion@1.1.7": {
+      "map": {
+        "balanced-match": "npm:balanced-match@0.4.2",
+        "concat-map": "npm:concat-map@0.0.1"
+      }
+    },
+    "npm:jspm-nodelibs-timers@0.2.1": {
+      "map": {
+        "timers-browserify": "npm:timers-browserify@1.4.2"
+      }
+    },
+    "npm:jspm-nodelibs-crypto@0.2.1": {
+      "map": {
+        "crypto-browserify": "npm:crypto-browserify@3.11.0"
+      }
+    },
+    "npm:jspm-nodelibs-os@0.2.1": {
+      "map": {
+        "os-browserify": "npm:os-browserify@0.2.1"
+      }
+    },
+    "npm:jspm-nodelibs-buffer@0.2.3": {
+      "map": {
+        "buffer": "npm:buffer@5.0.6"
+      }
+    },
+    "npm:jspm-nodelibs-stream@0.2.1": {
+      "map": {
+        "stream-browserify": "npm:stream-browserify@2.0.1"
+      }
+    },
+    "npm:jspm-nodelibs-string_decoder@0.2.1": {
+      "map": {
+        "string_decoder": "npm:string_decoder@0.10.31"
+      }
+    },
+    "npm:rxjs@5.3.1": {
       "map": {
         "symbol-observable": "npm:symbol-observable@1.0.4"
+      }
+    },
+    "npm:jspm-nodelibs-url@0.2.1": {
+      "map": {
+        "url": "npm:url@0.11.0"
+      }
+    },
+    "npm:buffer@5.0.6": {
+      "map": {
+        "ieee754": "npm:ieee754@1.1.8",
+        "base64-js": "npm:base64-js@1.2.0"
+      }
+    },
+    "npm:stream-http@2.7.0": {
+      "map": {
+        "inherits": "npm:inherits@2.0.3",
+        "readable-stream": "npm:readable-stream@2.2.9",
+        "builtin-status-codes": "npm:builtin-status-codes@3.0.0",
+        "to-arraybuffer": "npm:to-arraybuffer@1.0.1",
+        "xtend": "npm:xtend@4.0.1"
+      }
+    },
+    "npm:typescript@2.3.2": {
+      "map": {
+        "source-map-support": "npm:source-map-support@0.4.15"
+      }
+    },
+    "npm:source-map-support@0.4.15": {
+      "map": {
+        "source-map": "npm:source-map@0.5.6"
       }
     }
   }
